Implements movie details view
Adds the ability to view details for a selected movie.

This introduces a new state variable `showMovie` to track the currently
selected movie. When a movie item is clicked in the list, the
`handleShowMovie` function updates this state, triggering the display
of the `ShowMovie` component. The `ShowMovie` component renders basic
movie information. The Box component is updated to include a back button
to return to the movie list. Watched summary is also displayed only when
no movie is selected.